So, saving/loading the game caused every NPC to forget what they're doing... sometimes, maybe.
This is halfway fixed by setting their Unity GameObjects to inactive while loading, and setting active once loading is done. They were before executing their scripts before the game data was loaded, thought they weren't inside the building, and decided they couldn't work.

Now, though, they do continue their assigned work. But the work progress is not saved. This may not be fixable as it is the time to completion is only saved in the Script Component object, not the Entity Data which is what gets saved. But it is not as bad as workers going back to their office before going all the way back up to work.

Also included with this update is a log of debug spam while I work on this.